Add ServerModifyOxygen method for PlayerLife
This commit adds a new extension method to the PlayerLife class that allows setting the oxygen level of the player on the server side. The method takes one parameter: the new oxygen value to assign to the player. The method also invokes a network command to synchronize the UI data on the client side. The method uses reflection to access private fields and methods of the PlayerLife class. This method can be used by other classes or scripts that need to change the player’s oxygen level on the server side.

(Source: [https://github.com/SmartlyDressedGames/Unturned-3.x-Community/issues/3874#issuecomment-1594261459])